+++ scalr_tfc_migrate/constants.py
@@ -1,4 +1,4 @@
 """Migrator constants."""
 MAX_TERRAFORM_VERSION = "1.5.7"
 DEFAULT_MANAGEMENT_ENV_NAME = "scalr-admin"
-TFC_MIGRATOR_BACKEND_SECRETS_VARSET_NAME = "Scalr-Creds"  # created in TFC by init_backend_secrets; not migrated to Scalr
+TFC_MIGRATOR_DEFAULT_SECRETS_VARSET_NAME = "Scalr-Creds"  # created in TFC by init_backend_secrets; not migrated to Scalr